+++ tests/testthat/test-spatial.R
@@ -39,3 +39,30 @@ test_that("spatial don't duplicate samples", {
   expect_false(any(table(batch_names(index), batch_names(index2)) > 1))
 })
 
+test_that("Removing rows works", {
+  data(survey, package = "MASS")
+  set.seed(21829)
+  nas <- c(137, 70)
+  index <- design(survey[-nas, c("Sex", "Smoke", "Age")], size_subset = 50,
+                  iterations = 25)
+  index2 <- spatial(index, survey[-nas, c("Sex", "Smoke", "Age")],
+                    rows = LETTERS[1:9], remove_positions = "B",
+                    columns = 1:12, iterations = 25)
+  expect_lte(length(index2), 9*12)
+  expect_true(all(names(index2) %in% position_name(rows = LETTERS[1:9], 1:12)$name))
+  expect_false(any(table(batch_names(index), batch_names(index2)) > 1))
+})
+
+test_that("Removing columns works", {
+  data(survey, package = "MASS")
+  set.seed(21289)
+  nas <- c(137, 70)
+  index <- design(survey[-nas, c("Sex", "Smoke", "Age")], size_subset = 50,
+                  iterations = 25)
+  index2 <- spatial(index, survey[-nas, c("Sex", "Smoke", "Age")],
+                    rows = LETTERS[1:9], remove_positions =  "A",
+                    columns = 1:12, iterations = 25)
+  expect_lte(length(index2), 9*12)
+  expect_true(all(names(index2) %in% position_name(rows = LETTERS[1:9], 1:12)$name))
+  expect_false(any(table(batch_names(index), batch_names(index2)) > 1))
+})